Safari-609.1.11.1/LayoutTests/fast/shadow-dom/computed-style-display-none-expected.txt

Ensure tha slotted elements with display:none inherit their style via shadow tree

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(hostChild).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
